requirements.txt
streamlit>=1.32
python-dotenv>=1.0
tenacity>=8.2
google-generativeai>=0.3.0
PyPDF2>=3.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
google-genai=1.2.0